cleaned repo

+++ java-code/src/main/java/ch/ethz/systems/netbench/core/network/Link.java
@@ -19,7 +19,7 @@ public abstract class Link {
      *
      * @return  Link bandwidth in bit/ns
      */
-    public abstract long getBandwidthBitPerNs();
+    public abstract double getBandwidthBitPerNs();
 
     /**
      * Check whether the next transmission of a packet will fail.
