+++ java-code/src/main/java/ch/ethz/systems/netbench/ext/basic/EcnTailDropOutputPort.java
@@ -46,7 +46,7 @@ public class EcnTailDropOutputPort extends OutputPort {
 
                 // Add event when sending is finished
                 Simulator.registerEvent(new PacketDispatchedEvent(
-                        packet.getSizeBit() / getLink().getBandwidthBitPerNs(),
+                        (long)((double)packet.getSizeBit() / getLink().getBandwidthBitPerNs()),
                         packet,
                         this
                 ));
