+++ java-code/src/main/java/ch/ethz/systems/netbench/ext/basic/PerfectSimpleLinkGenerator.java
@@ -8,9 +8,9 @@ import ch.ethz.systems.netbench.core.run.infrastructure.LinkGenerator;
 public class PerfectSimpleLinkGenerator extends LinkGenerator {
 
     private final long delayNs;
-    private final long bandwidthBitPerNs;
+    private final double bandwidthBitPerNs;
 
-    public PerfectSimpleLinkGenerator(long delayNs, long bandwidthBitPerNs) {
+    public PerfectSimpleLinkGenerator(long delayNs, double bandwidthBitPerNs) {
         this.delayNs = delayNs;
         this.bandwidthBitPerNs = bandwidthBitPerNs;
         SimulationLogger.logInfo("Link", "PERFECT_SIMPLE_LINK(delayNs=" + delayNs + ", bandwidthBitPerNs=" + bandwidthBitPerNs + ")");
