+++ java-code/src/main/java/ch/ethz/systems/netbench/xpt/ports/FIFO/FIFOOutputPort.java
@@ -34,7 +34,7 @@ public class FIFOOutputPort extends OutputPort {
 
             // Add event when sending is finished
             Simulator.registerEvent(new PacketDispatchedEvent(
-                    packet.getSizeBit() / getLink().getBandwidthBitPerNs(),
+                    (long)((double)packet.getSizeBit() / getLink().getBandwidthBitPerNs()),
                     packet,
                     this
             ));
