+++ java-code/src/main/java/ch/ethz/systems/netbench/xpt/ports/PIFO_WFQ/WFQPIFOOutputPort.java
@@ -32,7 +32,7 @@ public class WFQPIFOOutputPort extends OutputPort {
 
             // Add event when sending is finished
             Simulator.registerEvent(new PacketDispatchedEvent(
-                    packet.getSizeBit() / getLink().getBandwidthBitPerNs(),
+                    (long)((double)packet.getSizeBit() / getLink().getBandwidthBitPerNs()),
                     packet,
                     this
             ));
