+++ java-code/src/main/java/ch/ethz/systems/netbench/xpt/ports/SPPIFO_WFQ/WFQSPPIFOOutputPort.java
@@ -31,7 +31,7 @@ public class WFQSPPIFOOutputPort extends OutputPort {
 
             // Add event when sending is finished
             Simulator.registerEvent(new PacketDispatchedEvent(
-                    packet.getSizeBit() / getLink().getBandwidthBitPerNs(),
+                    (long)((double)packet.getSizeBit() / getLink().getBandwidthBitPerNs()),
                     packet,
                     this
             ));
